fix(sdk): isolate Vertx code to prevent NoClassDefFoundError with JDK adapter
Fixes #7223 - When using the JDK HTTP adapter without Vertx dependencies
on the classpath, the SDK failed with NoClassDefFoundError because
RegistryClientRequestAdapterFactory had compile-time Vertx imports.

This change extracts all Vertx-specific code into a separate
VertxAdapterFactory class that is only loaded when Vertx adapter is
selected, leveraging JVM lazy class loading.

Changes:
- Create VertxAdapterFactory with all Vertx adapter creation logic
- Remove Vertx imports from RegistryClientRequestAdapterFactory
- Delegate to VertxAdapterFactory when VERTX adapter is selected
- Check HttpClosedException by class name instead of instanceof
